Refactor chunk generation and merging for forward and inverted indexes.
This all now takes place within a separate class that each index uses.
This eliminates a lot of nasty from the codebase---in particular, the
inverted_index class no longer has some member variables specifically for
merging chunks (but unused for all other purposes), and it consolidates
logic for this handling for both forward and inverted indexes.