Allow for newer versions of gcc to compile MeTA.
Tested with GCC 4.8. This is mostly a "hack" in the sense that it falls
back on some things that are in theory less performant for the sake of
getting it to work. It's not a hack in the sense that these features are
detected with CMake and eventually we can eliminate them when GCC catches
up with the standard...