Add factory method for creating filter chains.
Filters have been redesigned to contain a unique_ptr to their source
(references wouldn't work for this), and now all token_streams have a
method to set the content of the beginning of the filter chain. The chains
are not thread safe, so they will be created once at the beginning of each
thread that is involved in the tokenization process.