Fix bug with forward_index creation via uninverting.
The issue occurred when the inverted index had blank documents: empty
postings data entries were not being created in the uninverted forward
index in this case. This change requires that we now always write out the
primary key along with the postings data entry to intermediate chunks so
we can check it when univerting.