Add support for configuring language/country in icu_tokenizer.
The underlying ICU implementation is allowed to make changes to how
BreakIterator works depending on the Locale used in its construction. You
can now specify the language and country codes when creating an
icu_tokenizer if you know ahead of time what language you're going to be
dealing with.

By default, it uses the "en" "US" combination.