Improve memory utilization during indexing.
In memory chunks are now stored using the same on-disk compressed format
for the postings lists, but instead in in-memory byte buffers. These
buffers are organized into a custom insert-only linear probing hash set,
which allows us to have better control and measurement of memory
utilization.

This shaves off ~100 seconds when indexing Wikipedia on my local machine.
The entire dataset can be indexed without a single flush before the final
merge step with less than 8GB of ram.