Change analyzer to analyzer<T>.
This hasn't changed functionality much at all, yet, but prepares us for
changing the behavior of tokenize(doc) to depend on T.

This required some changing to how the analyzer factory worked to allow
for partial specialization in the factory functions (you can partially
specialize classes, but not functions). The change isn't too onerous, so
I'm not too worried about it.

The extern template and explicit template instantiation will need to be
explained in the documentation more, but I don't think it's too confusing:
essentially there are only ever two possible instantiations of your class,
so we don't actually want to header-ize everything and re-compile the
whole analyzer every time it's used. Instead, we force everything into a
library file, despite the fact that it's a template, because we can
explicitly instantiate our two cases at the end of the file.