Add a conditional compilation fix for GCC 4.8.1 ICE.
Nothing else fixes this that I've tried. GCC 4.8.1 just really hates
having a lambda function passed as the default argument to anything that
is a member function template. I don't know why those specific conditions
cause it, but eliminating any of them causes it to not ICE.

...but here we can't really remove any of those restrictions, so the
easiest thing is to just not pass a lambda for that compiler verison.