Add preliminary support for regression tasks.
This contains a bit of refactoring as well: libsvm format became a corpus
to allow configurable label handling, and that was finally refactored to
use the same factory pattern being applied to the other class hierarchies.
This also contains a fix for issue #86 since the format for the
breast-cancer corpus needed to be updated.