Ensure the doc label mapping is flushed before exiting make_index().
It was previously possible that the doc label mapping was not flushed to
disk before it was re-mapped by another instance of an index object. This
was only a bug when there were two copies of an index open at the same
time, one of which invoked create_index() and the other which invoked
load_index().