Fix a non-determinism in the one-vs-one reduction.
In certain cases, different compilers and standard libraries may pick
different orders to create the positive/negative class pairs in---for
example, (chinese, english) vs (english, chinese). This matters when using
SGD as the base classifier, as its termination depends on the loss
function which depends on which class is positive. To work around this, we
determinize the choice of positive/negative class pairs for the reduction
by sorting the class labels before creating the class pairs.