Refactor analyzer<T> to non-template analyzer.
This introduces the new analyzers::featurizer class, which is used as a
proxy to increment count data in a feature_map. analyzer::analyze() is now
a template function that indicates the feature value you wish to have
reported: uint64_t for inverted_index and double for forward_index. An
exception is thrown if an analyzer attempts to use a double feature value
when being used with uint64_t feature values.

This dramatically simplifies the boilerplate for adding new analyzers and
will make the documentation shorter.