Limit the number of writing threads during indexing.
This adds a parallel::semaphore class implementing a counting semaphore
with a mutex and condition variable. One has been added to
postings_inverter, allowing for a limit to be placed on the number of
concurrently writing threads. Systems with large numbers of threads (> 16)
and/or a slow disk were mostly impacted, especially if all threads never
flushed an intermediate chunk---common with large amounts of available RAM
budget---leading to every indexing thread hitting the disk at once.