Add calibration routine to learn::sgd_model.
This attempts to automatically set the learning rate based on the average
loss incurred on some small sample set. This helps *tremendously* for
regression, but seems to hurt classification performance by a small
amount. For that reason, it is enabled by default for the sgd regressor,
and disabled by default for the sgd classifier.