Begin initial work on GloVe training tool.
Refactor the existing embedding-coocur code to pull out the reusable
components into the meta::embeddings namespace.

Add the first step of GloVe training, which is to shuffle and partition
the coocurrence matrix into num_threads chunks, each of which can be
operated on in parallel.

Add initial skeleton of what interacting with the glove class might look
like.