allow full text storage in index as metadata field
If store-full-text = true (default false) in the corpus config, the string
metadata field "content" will be added. This is to simplify the creation of full
text metadata: the user doesn't have to duplicate their dataset in metadata.dat,
and metadata.dat will still be somewhat human-readable without large strings of
full text added.

This option has no affect for libsvm-corpus.