Add horrible hack workaround for remove_all on Windows.
On Windows, ::DeleteFile() only "marks a file for deletion on close" and
therefore "the file deletion does not occur until the last handle to the
file is closed". This means that, in some cases, the call to
remove_directory in remove_all can fail if another process (like the
Windows search indexer or a virus scanner) happens to still have one of
the files in this directory that we deleted above open when we attempt
to remove it.

As a workaround, we'll attempt to remove a directory a maximum of three
times, sleeping for 100ms more between each successive try. If removing
the directory fails three times in a row, we will throw an exception and
bail out. I really wish there were a better workaround than this, but I
can't come up with anything and we want remove_all() to have the same
(sane) semantics as it would have on Unix platforms.