Allow indexes to be loaded from read-only filesystems.
This changes the behavior of disk_vector<const T>, which was otherwise
unused in the toolkit. If disk_vector is instantiated with a const
datatype, it is opened in read only mode and mutating functions are
disabled. meta::index classes have been updated to use
disk_vector<const T> where appropriate to allow loading pre-existing
indexes in read-only mode.