Optimize memory usage in postings_inverter.
This adds two new classes: util::arena and util::arena_allocator. The
allocator supports a change of the postings_buffer internal byte buffer
to instead be a forward_list of fixed-size byte arrays, each of which
is allocated inside of a thread-local arena.

This allows postings_inverter::producer to pre-allocate nearly all of
the RAM that will be used during indexing and avoids the heap
fragmentation problems that could arise when not using jemalloc as the
underlying malloc implementation.