Fix GCC 7.1.1 issue with MAKE_NUMERIC_IDENTIFIER.
This is a *super bizarre* issue. This only appears if you use
MAKE_NUMERIC_IDENTIFIER instead of MAKE_NUMERIC_IDENTIFIER_UDL, and
only in some very specific circumstances. GCC seems to think that
numeric_identifier is not copyable, but only in some cases, some of the
time.

See this Gist for an example:
https://gist.github.com/skystrife/2e23a0f3fbc709b8d11d3a8c59c64e4a

Change pretty much any one thing in the example and the error goes
away. Make the constructor not a template? No error. Remove the
increment line in the constructor? No error. Define a user defined
literal after the definition of topic_id? No error. Put the
each_seen_event() function call into the constructor and remove the
non-template function? No error.

I have no idea what is going on, other than that GCC 7.1.1 is just
somehow very broken.