[learn]: switch to vanilla AdaGrad w/ calibration for sgd
The NAG algorithm has great performance on the housing regression
dataset which was the main reason it was initially implemented.
However, it actually reduces the performance of vanilla AdaGrad on
normal text classification datasets, which is the norm here since this
is a text analysis toolkit.

I also tried the Adam algorithm, but it did not perform as well
(robustly) as AdaGrad with learning rate calibration.

This commit also fixes a bug in the calibration algorithm when applied
to the perceptron loss (by just not calibrating for that loss function)
that prevented it from being enabled by default in the sgd-based
classifiers before.