[util]: Fix for string-based identifiers for Xcode 9.
There is apparently ambiguity for something like
static_cast<std::string>(var) if var is class_label on Xcode 9 because
there are two constructors for std::string that could apply: one from
const std::string& and one from std::string_view. Since class_label
would have conversion operators for both, the construction is ambiguous,
so force the string_view conversion operator to be explicit so it does
not participate in such a situation.